// packages/zotmeal-core/src/api/types/items.ts
export interface ArtworkItemData extends ItemBase {
  itemType: "artwork";
  creators: ReadonlyArray<CreatorBase<ArtworkCreatorType>>;
  title: string;
  abstractNote: string;
  artworkMedium: string;
  artworkSize: string;
  date: string;
  eventPlace: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Audio Recording */
export interface AudioRecordingItemData extends ItemBase {
  itemType: "audioRecording";
  creators: ReadonlyArray<CreatorBase<AudioRecordingCreatorType>>;
  title: string;
  abstractNote: string;
  audioRecordingFormat: string;
  seriesTitle: string;
  volume: string;
  numberOfVolumes: string;
  label: string;
  place: string;
  date: string;
  runningTime: string;
  ISBN: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Bill */
export interface BillItemData extends ItemBase {
  itemType: "bill";
  creators: ReadonlyArray<CreatorBase<BillCreatorType>>;
  title: string;
  abstractNote: string;
  billNumber: string;
  code: string;
  codeVolume: string;
  section: string;
  codePages: string;
  legislativeBody: string;
  session: string;
  history: string;
  date: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  shortTitle: string;
  language: string;
  rights: string;
  extra: string;
}

/** Blog Post */
export interface BlogPostItemData extends ItemBase {
  itemType: "blogPost";
  creators: ReadonlyArray<CreatorBase<BlogPostCreatorType>>;
  title: string;
  abstractNote: string;
  blogTitle: string;
  websiteType: string;
  date: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  ISSN: string;
  shortTitle: string;
  language: string;
  rights: string;
  extra: string;
}

/** Book */
export interface BookItemData extends ItemBase {
  itemType: "book";
  creators: ReadonlyArray<CreatorBase<BookCreatorType>>;
  title: string;
  abstractNote: string;
  series: string;
  seriesNumber: string;
  volume: string;
  numberOfVolumes: string;
  edition: string;
  date: string;
  publisher: string;
  place: string;
  originalDate: string;
  originalPublisher: string;
  originalPlace: string;
  format: string;
  numPages: string;
  ISBN: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  ISSN: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Book Section */
export interface BookSectionItemData extends ItemBase {
  itemType: "bookSection";
  creators: ReadonlyArray<CreatorBase<BookSectionCreatorType>>;
  title: string;
  abstractNote: string;
  bookTitle: string;
  series: string;
  seriesNumber: string;
  volume: string;
  numberOfVolumes: string;
  edition: string;
  date: string;
  publisher: string;
  place: string;
  originalDate: string;
  originalPublisher: string;
  originalPlace: string;
  format: string;
  pages: string;
  ISBN: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  ISSN: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Case */
export interface CaseItemData extends ItemBase {
  itemType: "case";
  creators: ReadonlyArray<CreatorBase<CaseCreatorType>>;
  caseName: string;
  abstractNote: string;
  court: string;
  dateDecided: string;
  docketNumber: string;
  reporter: string;
  reporterVolume: string;
  firstPage: string;
  history: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  shortTitle: string;
  language: string;
  rights: string;
  extra: string;
}

/** Conference Paper */
export interface ConferencePaperItemData extends ItemBase {
  itemType: "conferencePaper";
  creators: ReadonlyArray<CreatorBase<ConferencePaperCreatorType>>;
  title: string;
  abstractNote: string;
  proceedingsTitle: string;
  conferenceName: string;
  publisher: string;
  place: string;
  date: string;
  eventPlace: string;
  volume: string;
  issue: string;
  numberOfVolumes: string;
  pages: string;
  series: string;
  seriesNumber: string;
  DOI: string;
  ISBN: string;
  citationKey: string;
  url: string;
  accessDate: string;
  ISSN: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Dataset */
export interface DatasetItemData extends ItemBase {
  itemType: "dataset";
  creators: ReadonlyArray<CreatorBase<DatasetCreatorType>>;
  title: string;
  abstractNote: string;
  identifier: string;
  type: string;
  versionNumber: string;
  date: string;
  repository: string;
  repositoryLocation: string;
  format: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Dictionary Entry */
export interface DictionaryEntryItemData extends ItemBase {
  itemType: "dictionaryEntry";
  creators: ReadonlyArray<CreatorBase<DictionaryEntryCreatorType>>;
  title: string;
  abstractNote: string;
  dictionaryTitle: string;
  series: string;
  seriesNumber: string;
  volume: string;
  numberOfVolumes: string;
  edition: string;
  date: string;
  publisher: string;
  place: string;
  pages: string;
  ISBN: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Document */
export interface DocumentItemData extends ItemBase {
  itemType: "document";
  creators: ReadonlyArray<CreatorBase<DocumentCreatorType>>;
  title: string;
  abstractNote: string;
  type: string;
  date: string;
  publisher: string;
  place: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** E-mail */
export interface EmailItemData extends ItemBase {
  itemType: "email";
  creators: ReadonlyArray<CreatorBase<EmailCreatorType>>;
  subject: string;
  abstractNote: string;
  date: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  shortTitle: string;
  language: string;
  rights: string;
  extra: string;
}

/** Encyclopedia Article */
export interface EncyclopediaArticleItemData extends ItemBase {
  itemType: "encyclopediaArticle";
  creators: ReadonlyArray<CreatorBase<EncyclopediaArticleCreatorType>>;
  title: string;
  abstractNote: string;
  encyclopediaTitle: string;
  series: string;
  seriesNumber: string;
  volume: string;
  numberOfVolumes: string;
  edition: string;
  date: string;
  publisher: string;
  place: string;
  pages: string;
  ISBN: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Film */
export interface FilmItemData extends ItemBase {
  itemType: "film";
  creators: ReadonlyArray<CreatorBase<FilmCreatorType>>;
  title: string;
  abstractNote: string;
  distributor: string;
  place: string;
  date: string;
  genre: string;
  videoRecordingFormat: string;
  runningTime: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Forum Post */
export interface ForumPostItemData extends ItemBase {
  itemType: "forumPost";
  creators: ReadonlyArray<CreatorBase<ForumPostCreatorType>>;
  title: string;
  abstractNote: string;
  forumTitle: string;
  postType: string;
  date: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  shortTitle: string;
  language: string;
  rights: string;
  extra: string;
}

/** Hearing */
export interface HearingItemData extends ItemBase {
  itemType: "hearing";
  creators: ReadonlyArray<CreatorBase<HearingCreatorType>>;
  title: string;
  abstractNote: string;
  committee: string;
  publisher: string;
  numberOfVolumes: string;
  documentNumber: string;
  pages: string;
  legislativeBody: string;
  session: string;
  history: string;
  date: string;
  place: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  shortTitle: string;
  language: string;
  rights: string;
  extra: string;
}

/** Instant Message */
export interface InstantMessageItemData extends ItemBase {
  itemType: "instantMessage";
  creators: ReadonlyArray<CreatorBase<InstantMessageCreatorType>>;
  title: string;
  abstractNote: string;
  date: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  shortTitle: string;
  language: string;
  rights: string;
  extra: string;
}

/** Interview */
export interface InterviewItemData extends ItemBase {
  itemType: "interview";
  creators: ReadonlyArray<CreatorBase<InterviewCreatorType>>;
  title: string;
  abstractNote: string;
  interviewMedium: string;
  date: string;
  publisher: string;
  place: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}

/** Journal Article */
export interface JournalArticleItemData extends ItemBase {
  itemType: "journalArticle";
  creators: ReadonlyArray<CreatorBase<JournalArticleCreatorType>>;
  title: string;
  abstractNote: string;
  publicationTitle: string;
  publisher: string;
  place: string;
  date: string;
  volume: string;
  issue: string;
  section: string;
  partNumber: string;
  partTitle: string;
  pages: string;
  series: string;
  seriesTitle: string;
  seriesText: string;
  journalAbbreviation: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  PMID: string;
  PMCID: string;
  ISSN: string;
  archive: string;
  archiveLocation: string;
  shortTitle: string;
  language: string;
  libraryCatalog: string;
  callNumber: string;
  rights: string;
  extra: string;
}
// ...
export interface WebpageItemData extends ItemBase {
  itemType: "webpage";
  creators: ReadonlyArray<CreatorBase<WebpageCreatorType>>;
  title: string;
  abstractNote: string;
  websiteTitle: string;
  websiteType: string;
  date: string;
  publisher: string;
  place: string;
  DOI: string;
  citationKey: string;
  url: string;
  accessDate: string;
  shortTitle: string;
  language: string;
  rights: string;
  extra: string;
}
// ...
export type ItemData =
  | ArtworkItemData
  | AudioRecordingItemData
  | BillItemData
  | BlogPostItemData
  | BookItemData
  | BookSectionItemData
  | CaseItemData
  | ConferencePaperItemData
  | DatasetItemData
  | DictionaryEntryItemData
  | DocumentItemData
  | EmailItemData
  | EncyclopediaArticleItemData
  | FilmItemData
  | ForumPostItemData
  | HearingItemData
  | InstantMessageItemData
  | InterviewItemData
  | JournalArticleItemData
  | LetterItemData
  | MagazineArticleItemData
  | ManuscriptItemData
  | MapItemData
  | NewspaperArticleItemData
  | PatentItemData
  | PodcastItemData
  | PreprintItemData
  | PresentationItemData
  | RadioBroadcastItemData
  | ReportItemData
  | ComputerProgramItemData
  | StandardItemData
  | StatuteItemData
  | TvBroadcastItemData
  | ThesisItemData
  | VideoRecordingItemData
  | WebpageItemData
  | AttachmentItemData
  | NoteItemData
  | AnnotationItemData;